marketmind_backend/user-service/requirements.txt
Flask
Werkzeug
pymongo
pyjwt
flask-cors
requests-oauthlib
python-dotenv

